English: Add letters used in semitic transliteration.
Add additional characters to the popups on the qwerty-with-symbols
layout which are helpful in latin transliteration of semitic languages.
These are not exhaustive, since there are multiple schemes and I'm
only familiar with a few ancient languages.

Emphatic variants: ṣṭṯḏ
Back fricatives: ḥḫẖ

Added half ring character to the quote pop-up for transcribing
alef/hamza and ayin. Also added the turned comma sometimes used
for ayin, and for the glottal stop in polynesian languages. It's
also available on the shared quotes key by the spacebar.

Added half-square brackets used for marking partial restorations
from damaged manuscripts.