fix to #2252 added fe/male ordinals to pt keyboard
The ordinals ("first" can be written as "1º" or "1ª" depending on gender) are in very common use in Portuguese, even having a dedicated key on Portuguese keyboards.
In software keyboards they are usually implemented as alternative symbols to A and O.
This was the only file i found where these changes made sense.
This is a proposed fix to https://github.com/AnySoftKeyboard/AnySoftKeyboard/issues/2252